work with label github-pending rather than inbox
i just set up a single filter in gmail to put all github notifications in to
a folder named Github/Pending and then made the changes in this commit to
work off of that rather than my inbox.

it's just hacked in to the labler.gs since i'm not familar with coffee
script and i was in the middle of all day onboarding when i did it. it
should be pretty simple to make it an alternate mode rather than hard coded
this way.