Unmute first, then unsubscribe
Update muter to unmute the Gmail thread first before making the HTTP
request to unsubscribe from the GitHub thread. This way, if the HTTP
request fails, at least the Gmail thread will appear in the inbox if you
get a new message that @mentions you or a team that you're on.

Prior to this change, it was possible for the HTTP request to fail and
for the Gmail thread to get stuck in a muted state, preventing you from
seeing any new messages that @mention you or a team that you're on.

xref: https://github.com/mastahyeti/OctoGAS/issues/17